+++ src/all/komga/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.27
+
+Minimum Komga version required: `0.97.0`
+
+### Fix
+
+* filter series by read or in progress
+
 ## 1.2.26
 
 Minimum Komga version required: `0.87.4`